refactor: extract preview components from CardFactoryModal.tsx (#8608) [part 3]
108-line code movement out of CardFactoryModal.tsx (now 1556->1448 lines)
into a new cardFactoryPreviews.tsx. Zero behavior change; purely moves the
three presentational helpers (TemplateDropdown, T1Preview, T2Preview) that
render template pickers and AI generation result previews.

Follow-up to #8834 and #8838; further extractions to come.
